refactored code to be compilable
